Avoid using a local window for splitting
The splitting code in bupsplit.c/_hashsplit.c was written as if it had
to handle the case where bytes were fed one by-one, but it actually gets
the whole data chunk: there's no need to copy to a local window, when we
have the original data available: we can simply use the passed data to
fetch the byte that is leaving the current window. This should actually
be a fair bit faster.

Additionally, try to use __builtin_ctz for getting extra digest bits if
it is available.